+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 ## v2.8.0
 * Added `d!coinfliptimes`
     - Flips a coin the amount of times you want, and shows you the outcome.
+* Added `d!link`
+    - Shorten a link with doob.link!
+        - 6 links for regular people
+        - 12 links for [Patrons](https://patreon.com/doobdev/)
+* Added /link
 
 ## v2.7.2
 * Added /support